fix: ask for a full game packet on desync
