+++ server/include/Player.hpp
@@ -9,6 +9,7 @@
 #include "Tetris.hpp"
 #include "network/APacket.hpp"
 #include "network/PacketHandler.hpp"
+#include "network/packets/FullGameRequestPacket.hpp"
 #include "network/packets/GameActionPacket.hpp"
 #include <deque>
 #include <enet/enet.h>
@@ -32,6 +33,7 @@ namespace tetriq {
             void sendPacket(const APacket &packet);
 
             bool handle(GameActionPacket &packet);
+            bool handle(FullGameRequestPacket &packet);
 
         private:
             const uint64_t _network_id;
